Convert table wrapper to package component and create readme docs
