+++ examples/simple-table/src/index.jsx
@@ -3,7 +3,8 @@ import ReactDOM from "react-dom/client";
 import "./styles/theme.css";
 import App from "./App";
 import { OfflineStorageWrapper } from "./packages/contexts/OfflineStorageWrapper";
-import { TableWrapper } from "./contexts/TableWrapper";
+import { TableWrapper } from "./packages/contexts/TableWrapper";
+import { createColumnHelper } from "@tanstack/react-table";
 
 const offlineStorageConfig = {
   type: "indexedDB",
@@ -14,12 +15,37 @@ const offlineStorageConfig = {
   },
 };
 
+const columnHelper = createColumnHelper();
+
+const columnMap = [
+  columnHelper.accessor("isDraft", {
+    cell: (info) => (info.getValue() ? "Draft " : "Submitted"),
+    header: () => <span>Status</span>,
+  }),
+  columnHelper.accessor("uuid", {
+    cell: (info) => info.getValue(),
+    header: () => <span>Id</span>,
+  }),
+  columnHelper.accessor("species", {
+    cell: (info) => info.getValue(),
+    header: () => <span>Species</span>,
+  }),
+  columnHelper.accessor("numberOfFish", {
+    cell: (info) => info.getValue(),
+    header: () => <span>Amount Caught</span>,
+  }),
+  columnHelper.accessor("computedPrice", {
+    cell: (info) => info.getValue(),
+    header: () => <span>Price</span>,
+  }),
+];
+
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
 root.render(
   <React.StrictMode>
     <OfflineStorageWrapper config={offlineStorageConfig}>
-      <TableWrapper>
+      <TableWrapper columnMap={columnMap} pageSize={10}>
         <App />
       </TableWrapper>
     </OfflineStorageWrapper>
